mcp_agent/requirements.txt
google-adk>=1.0.0
sseclient-py>=1.7.2
requests>=2.31.0